pyproject.toml
[project]
name = "asair-ai-sandbox"
version = "0.1.0"
description = "Dify 提示词 & Chatflow 沙盒测试工具"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "jinja2>=3.1.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "black>=24.0",
    "ruff>=0.4",
    "pre-commit>=3.0",
]

[project.scripts]
sandbox = "sandbox.cli:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
